+++ src/main/java/org/opentripplanner/netex/mapping/NetexMapper.java
@@ -303,6 +303,7 @@ public class NetexMapper {
       idFactory,
       currentNetexIndex.getQuayById(),
       tariffZoneMapper,
+      transitBuilder.stopModelBuilder(),
       zoneId,
       issueStore,
       noTransfersOnIsolatedStops