+++ src/main/java/org/opentripplanner/raptor/rangeraptor/multicriteria/configure/McRangeRaptorConfig.java
@@ -70,6 +70,7 @@ public class McRangeRaptorConfig<T extends RaptorTripSchedule> {
     return new McStopArrivals<>(
       context.nStops(),
       context.egressPaths(),
+      context.accessPaths(),
       createDestinationArrivalPaths(),
       context.debugFactory()
     );