+++ src/main/java/org/opentripplanner/standalone/server/DefaultServerRequestContext.java
@@ -1,7 +1,6 @@
 package org.opentripplanner.standalone.server;
 
 import io.micrometer.core.instrument.MeterRegistry;
-import java.time.Duration;
 import java.util.Locale;
 import javax.annotation.Nullable;
 import org.opentripplanner.astar.spi.TraverseVisitor;
@@ -33,7 +32,6 @@ public class DefaultServerRequestContext implements OtpServerRequestContext {
   private final Graph graph;
   private final TransitService transitService;
   private final TransitRoutingConfig transitRoutingConfig;
-  private final Duration streetRoutingTimeout;
   private final RouteRequest routeRequestDefaults;
   private final MeterRegistry meterRegistry;
   private final RaptorConfig<TripSchedule> raptorConfig;
@@ -53,7 +51,6 @@ public class DefaultServerRequestContext implements OtpServerRequestContext {
     Graph graph,
     TransitService transitService,
     TransitRoutingConfig transitRoutingConfig,
-    Duration streetRoutingTimeout,
     RouteRequest routeRequestDefaults,
     MeterRegistry meterRegistry,
     RaptorConfig<TripSchedule> raptorConfig,
@@ -69,7 +66,6 @@ public class DefaultServerRequestContext implements OtpServerRequestContext {
     this.graph = graph;
     this.transitService = transitService;
     this.transitRoutingConfig = transitRoutingConfig;
-    this.streetRoutingTimeout = streetRoutingTimeout;
     this.meterRegistry = meterRegistry;
     this.raptorConfig = raptorConfig;
     this.requestLogger = requestLogger;
@@ -89,7 +85,6 @@ public class DefaultServerRequestContext implements OtpServerRequestContext {
   public static DefaultServerRequestContext create(
     TransitRoutingConfig transitRoutingConfig,
     RouteRequest routeRequestDefaults,
-    Duration streetRoutingTimeout,
     RaptorConfig<TripSchedule> raptorConfig,
     Graph graph,
     TransitService transitService,
@@ -106,7 +101,6 @@ public class DefaultServerRequestContext implements OtpServerRequestContext {
       graph,
       transitService,
       transitRoutingConfig,
-      streetRoutingTimeout,
       routeRequestDefaults,
       meterRegistry,
       raptorConfig,
@@ -181,11 +175,6 @@ public class DefaultServerRequestContext implements OtpServerRequestContext {
   @Override
   public RaptorTuningParameters raptorTuningParameters() {
     return transitRoutingConfig;
-  }
-
-  @Override
-  public Duration streetRoutingTimeout() {
-    return streetRoutingTimeout;
   }
 
   @Override