+++ src/test/java/org/opentripplanner/raptor/rangeraptor/multicriteria/StopArrivalStateParetoSetTest.java
@@ -1,9 +1,16 @@
 package org.opentripplanner.raptor.rangeraptor.multicriteria;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.opentripplanner.raptor.rangeraptor.multicriteria.StopArrivalParetoSet.createEgressStopArrivalSet;
+import static org.opentripplanner.raptor.rangeraptor.multicriteria.StopArrivalParetoSet.createStopArrivalSet;
 
 import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.opentripplanner.raptor._data.transit.TestAccessEgress;
 import org.opentripplanner.raptor._data.transit.TestTransfer;
 import org.opentripplanner.raptor._data.transit.TestTripSchedule;
@@ -42,19 +49,13 @@ public class StopArrivalStateParetoSetTest {
     5,
     BASE_COST
   );
-  private final StopArrivalParetoSet<RaptorTripSchedule> subject = new StopArrivalParetoSet<>(null);
+
   private static final AbstractStopArrival<RaptorTripSchedule> TRANSIT_L1 = newTransitStopState(
     ROUND_1,
     998,
     10,
     BASE_COST
   );
-
-  @Test
-  public void addOneElementToSet() {
-    subject.add(newAccessStopState(STOP_1, 10, ANY));
-    assertStopsInSet(STOP_1);
-  }
 
   private static final AbstractStopArrival<RaptorTripSchedule> TRANSIT_L2 = newTransitStopState(
     ROUND_2,
@@ -63,64 +64,115 @@ public class StopArrivalStateParetoSetTest {
     BASE_COST
   );
 
-  @Test
-  public void testTimeDominance() {
+  private final StopArrivalParetoSet<RaptorTripSchedule> subject = createStopArrivalSet(null);
+
+  private static Stream<Arguments> testCases() {
+    return Stream.of(
+      Arguments.of("Stop Arrival - regular", createStopArrivalSet(null)),
+      Arguments.of("Stop Arrival - w/egress", createEgressStopArrivalSet(List.of(), null, null))
+    );
+  }
+
+  @ParameterizedTest(name = "{0}")
+  @MethodSource("testCases")
+  public void addOneElementToSet(
+    String testCaseName,
+    StopArrivalParetoSet<RaptorTripSchedule> subject
+  ) {
+    subject.add(newAccessStopState(STOP_1, 10, ANY));
+    assertStopsInSet(subject, STOP_1);
+  }
+
+  @ParameterizedTest(name = "{0}")
+  @MethodSource("testCases")
+  public void testTimeDominance(
+    String testCaseName,
+    StopArrivalParetoSet<RaptorTripSchedule> subject
+  ) {
     subject.add(newAccessStopState(STOP_1, 10, ANY));
     subject.add(newAccessStopState(STOP_2, 9, ANY));
     subject.add(newAccessStopState(STOP_3, 9, ANY));
     subject.add(newAccessStopState(STOP_4, 11, ANY));
-    assertStopsInSet(STOP_2);
+    assertStopsInSet(subject, STOP_2);
   }
 
-  @Test
-  public void testRoundDominance() {
+  @ParameterizedTest(name = "{0}")
+  @MethodSource("testCases")
+  public void testRoundDominance(
+    String testCaseName,
+    StopArrivalParetoSet<RaptorTripSchedule> subject
+  ) {
     subject.add(newTransferStopState(ROUND_1, STOP_1, 10, ANY));
     subject.add(newTransferStopState(ROUND_2, STOP_2, 10, ANY));
-    assertStopsInSet(STOP_1);
+    assertStopsInSet(subject, STOP_1);
   }
 
-  @Test
-  public void testCostDominance() {
+  @ParameterizedTest(name = "{0}")
+  @MethodSource("testCases")
+  public void testCostDominance(
+    String testCaseName,
+    StopArrivalParetoSet<RaptorTripSchedule> subject
+  ) {
     subject.add(newTransferStopState(ROUND_1, STOP_1, ANY, 20));
     subject.add(newTransferStopState(ROUND_1, STOP_2, ANY, 10));
-    assertStopsInSet(STOP_2);
+    assertStopsInSet(subject, STOP_2);
   }
 
-  @Test
-  public void testRoundAndTimeDominance() {
+  @ParameterizedTest(name = "{0}")
+  @MethodSource("testCases")
+  public void testRoundAndTimeDominance(
+    String testCaseName,
+    StopArrivalParetoSet<RaptorTripSchedule> subject
+  ) {
     subject.add(newTransferStopState(ROUND_1, STOP_1, 10, ANY));
     subject.add(newTransferStopState(ROUND_1, STOP_2, 8, ANY));
 
-    assertStopsInSet(STOP_2);
+    assertStopsInSet(subject, STOP_2);
 
     subject.add(newTransferStopState(ROUND_2, STOP_3, 8, ANY));
 
-    assertStopsInSet(STOP_2);
+    assertStopsInSet(subject, STOP_2);
 
     subject.add(newTransferStopState(ROUND_2, STOP_4, 7, ANY));
 
-    assertStopsInSet(STOP_2, STOP_4);
+    assertStopsInSet(subject, STOP_2, STOP_4);
 
     subject.add(newTransferStopState(ROUND_3, STOP_5, 6, ANY));
 
-    assertStopsInSet(STOP_2, STOP_4, STOP_5);
+    assertStopsInSet(subject, STOP_2, STOP_4, STOP_5);
 
     subject.add(newTransferStopState(ROUND_3, STOP_6, 6, ANY));
 
-    assertStopsInSet(STOP_2, STOP_4, STOP_5);
+    assertStopsInSet(subject, STOP_2, STOP_4, STOP_5);
   }
 
   /**
-   * During the same round transfers should not dominate transits, but this is handled by the worker
+   * During the same round, transfers should not dominate transits. This is handled by the worker
    * state (2-phase transfer calculation), not by the pareto-set. Using the pareto-set for this
    * would cause unnecessary exploration in the following round.
    */
   @Test
   public void testTransitAndTransferDoesNotAffectDominance() {
+    var subject = createStopArrivalSet(null);
     subject.add(newAccessStopState(STOP_1, 20, ANY));
     subject.add(newTransitStopState(ROUND_1, STOP_2, 10, ANY));
     subject.add(newTransferStopState(ROUND_1, STOP_4, 8, ANY));
-    assertStopsInSet(STOP_1, STOP_4);
+    assertStopsInSet(subject, STOP_1, STOP_4);
+  }
+
+  /**
+   * For arrivals arriving on-a-ride (transit and flex) we need to keep the arrival. These
+   * arrivals can continue with an egress that start with a walking leg.
+   *
+   * @see #testTransitAndTransferDoesNotAffectDominance
+   */
+  @Test
+  public void testTransitAndTransferDoesAffectDominanceForStopArrivalsWithEgress() {
+    var subject = createEgressStopArrivalSet(List.of(), null, null);
+    subject.add(newAccessStopState(STOP_1, 20, ANY));
+    subject.add(newTransitStopState(ROUND_1, STOP_2, 10, ANY));
+    subject.add(newTransferStopState(ROUND_1, STOP_4, 8, ANY));
+    assertStopsInSet(subject, STOP_1, STOP_2, STOP_4);
   }
 
   private static AccessStopArrival<RaptorTripSchedule> newAccessStopState(
@@ -171,7 +223,10 @@ public class StopArrivalStateParetoSetTest {
     }
   }
 
-  private void assertStopsInSet(int... expStopIndexes) {
+  private void assertStopsInSet(
+    StopArrivalParetoSet<RaptorTripSchedule> subject,
+    int... expStopIndexes
+  ) {
     int[] result = subject.stream().mapToInt(AbstractStopArrival::stop).sorted().toArray();
     assertEquals(Arrays.toString(expStopIndexes), Arrays.toString(result), "Stop indexes");
   }